Delete Cached and move LazyMacros to version specific sources
